// API_Client/client/models/revenue-over-time-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Shop API
 * REST API for the online shop backend
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { RevenueBucketDto } from './revenue-bucket-dto';

export interface RevenueOverTimeDto {
    'buckets': Array<RevenueBucketDto>;
    /**
     * მოთხოვნილი პერიოდის ჯამური შემოსავალი
     */
    'totalRevenue': number;
    /**
     * იმავე ხანგრძლივობის წინა პერიოდის ჯამური შემოსავალი
     */
    'previousPeriodRevenue': number;
    /**
     * ცვლილება პროცენტებში წინა პერიოდთან შედარებით — null, თუ წინა პერიოდის შემოსავალი 0 იყო
     */
    'changePercent': number | null;
}